+++ com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeReferenceVolume.cs
@@ -547,8 +547,10 @@ namespace UnityEngine.Experimental.Rendering
 
         bool m_NeedLoadAsset = false;
         bool m_ProbeReferenceVolumeInit = false;
+        bool m_EnabledBySRP = false;
 
         internal bool isInitialized => m_ProbeReferenceVolumeInit;
+        internal bool enabledBySRP => m_EnabledBySRP;
 
         struct InitInfo
         {
@@ -606,6 +608,7 @@ namespace UnityEngine.Experimental.Rendering
 
         /// <summary>
         /// Initialize the Probe Volume system
+        /// </summary>
         /// <param name="parameters">Initialization parameters.</param>
         public void Initialize(in ProbeVolumeSystemParameters parameters)
         {
@@ -628,8 +631,18 @@ namespace UnityEngine.Experimental.Rendering
                 UnityEditor.SceneManagement.EditorSceneManager.sceneSaved += sceneData.OnSceneSaved;
             }
 #endif
+            m_EnabledBySRP = true;
         }
 
+        /// <summary>
+        /// Communicate to the Probe Volume system whether the SRP enables Probe Volume.
+        /// It is important to keep in mind that this is not used by the system for anything else but book-keeping,
+        /// the SRP is still responsible to disable anything Probe volume related on SRP side.
+        /// </summary>
+        public void SetEnableStateFromSRP(bool srpEnablesPV)
+        {
+            m_EnabledBySRP = srpEnablesPV;
+        }
 
         // This is used for steps such as dilation that require the maximum order allowed to be loaded at all times. Should really never be used as a general purpose function.
         internal void ForceSHBand(ProbeVolumeSHBands shBands)
@@ -806,6 +819,12 @@ namespace UnityEngine.Experimental.Rendering
             }
         }
 
+        internal void SetMinBrickAndMaxSubdiv(float minBrickSize, int maxSubdiv)
+        {
+            SetTRS(Vector3.zero, Quaternion.identity, minBrickSize);
+            SetMaxSubdivision(maxSubdiv);
+        }
+
         void LoadAsset(ProbeVolumeAsset asset)
         {
             if (asset.Version != (int)ProbeVolumeAsset.AssetVersion.Current)
@@ -817,9 +836,7 @@ namespace UnityEngine.Experimental.Rendering
             var path = asset.GetSerializedFullPath();
 
             // Load info coming originally from profile
-            SetTRS(Vector3.zero, Quaternion.identity, asset.minBrickSize);
-            SetMaxSubdivision(asset.maxSubdivision);
-
+            SetMinBrickAndMaxSubdiv(asset.minBrickSize, asset.maxSubdivision);
 
             for (int i = 0; i < asset.cells.Count; ++i)
             {